+++ apis/python/src/tiledbsoma/_sparse_nd_array.py
@@ -6,14 +6,18 @@
 
 from __future__ import annotations
 
-from typing import Sequence, cast
+from typing import (
+    TYPE_CHECKING,
+    Sequence,
+    cast,
+)
 
 import numpy as np
 import pyarrow as pa
 import somacore
 from somacore import options
 from somacore.options import PlatformConfig
-from typing_extensions import Self
+from typing_extensions import Self, Unpack
 
 from . import _util
 
@@ -21,6 +25,7 @@ from . import _util
 from . import pytiledbsoma as clib
 from ._arrow_types import pyarrow_to_carrow_type
 from ._common_nd_array import NDArray
+from ._dask.util import SOMADaskConfig
 from ._exception import SOMAError, map_exception_for_create
 from ._read_iters import (
     BlockwiseScipyReadIter,
@@ -31,6 +36,7 @@ from ._read_iters import (
 )
 from ._tdb_handles import SparseNDArrayWrapper
 from ._types import NTuple, OpenTimestamp
+from ._util import from_clib_result_order
 from .options._soma_tiledb_context import (
     SOMATileDBContext,
     _validate_soma_tiledb_context,
@@ -39,6 +45,13 @@ from .options._tiledb_create_write_options import (
     TileDBCreateOptions,
     TileDBWriteOptions,
 )
+
+if TYPE_CHECKING:
+    try:
+        import dask.array as da
+    except ImportError:
+        pass
+
 
 _UNBATCHED = options.BatchSize()
 
@@ -446,6 +459,27 @@ class SparseNDArrayRead(_SparseNDArrayReadBase):
     Lifecycle:
         Maturing.
     """
+
+    def dask_array(
+        self,
+        **config: Unpack[SOMADaskConfig],
+    ) -> "da.Array":
+        """Load a TileDB-SOMA X layer as a Dask array.
+
+        The returned Array is effectively read-only; writes to it will not be persisted back to the
+        underlying TileDB-SOMA SparseNDArray.
+
+        Lifecycle: experimental
+        """
+        from tiledbsoma._dask.load import load_daskarray
+
+        return load_daskarray(
+            layer=self.array,
+            coords=self.coords,
+            result_order=from_clib_result_order(self.result_order),
+            platform_config=self.platform_config,
+            **config,
+        )
 
     def coos(self, shape: NTuple | None = None) -> SparseCOOTensorReadIter:
         """Returns an iterator of